fix type

+++ src/CommitmentsV2/SFA.DAS.CommitmentsV2.Api.Types/Responses/GetApprenticeshipUpdatesResponse.cs
@@ -19,7 +19,7 @@ namespace SFA.DAS.CommitmentsV2.Api.Types.Responses
             public string Email { get; set; }
             public DeliveryModel? DeliveryModel { get; set; }
             public DateTime? EmploymentEndDate { get; set; }
-            public decimal? EmploymentPrice { get; set; }
+            public int? EmploymentPrice { get; set; }
             public ProgrammeType? TrainingType { get; set; }
             public string TrainingCode { get; set; }
             public string Version { get; set; }
